requirements-dev.txt
pytest==8.3.2
PyYAML==6.0.2
